Avoid failures due to non-deterministic binding ordering
