+++ crates/ruff_linter/resources/test/fixtures/refurb/FURB163.py
@@ -1,9 +1,6 @@
 import math
 
-from math import e as special_e
-from math import log as special_log
-
-# Errors.
+# Errors
 math.log(1, 2)
 math.log(1, 10)
 math.log(1, math.e)
@@ -11,15 +8,10 @@ foo = ...
 math.log(foo, 2)
 math.log(foo, 10)
 math.log(foo, math.e)
-math.log(1, special_e)
-special_log(1, 2)
-special_log(1, 10)
-special_log(1, math.e)
-special_log(1, special_e)
 math.log(1, 2.0)
 math.log(1, 10.0)
 
-# Ok.
+# OK
 math.log2(1)
 math.log10(1)
 math.log(1)
@@ -39,6 +31,7 @@ math.log2(1, 10)  # math.log2 takes only one argument.
 math.log10(1, 2)  # math.log10 takes only one argument.
 
 math.log(1, base=2)  # math.log does not accept keyword arguments.
+
 
 def log(*args):
     print(f"Logging: {args}")
